python-backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
google-generativeai==0.8.6
chromadb==1.5.5
pypdf==6.9.2
sentence-transformers==5.3.0
python-dotenv==1.1.1
python-multipart==0.0.22
requests==2.32.5
pydantic==2.11.7
httpx==0.28.1
starlette==0.52.1
